chore: restructure database client imports and remove obsolete migration files
